Refactored files into the correct folder locations
