+++ test/v1/models/requestData/CreateCisDeductionsRequestModelSpec.scala
@@ -18,6 +18,7 @@ package v1.models.requestData
 
 import play.api.libs.json.{JsError, JsSuccess, JsValue, Json}
 import support.UnitSpec
+import v1.models.request.{CreateCisDeductionsRequestModel, PeriodDetails}
 
 class CreateCisDeductionsRequestModelSpec extends UnitSpec {
